# docs/relay/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UAM Relay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/blocklist": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Blocked",
        "description": "List all blocklist entries.",
        "operationId": "list_blocked_api_v1_admin_blocklist_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlocklistListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Add Blocked",
        "description": "Add a pattern to the blocklist.",
        "operationId": "add_blocked_api_v1_admin_blocklist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlocklistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Add Blocked Api V1 Admin Blocklist Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/blocklist/{pattern}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Remove Blocked",
        "description": "Remove a pattern from the blocklist.",
        "operationId": "remove_blocked_api_v1_admin_blocklist__pattern__delete",
        "parameters": [
          {
            "name": "pattern",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pattern"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove Blocked Api V1 Admin Blocklist  Pattern  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/allowlist": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Allowed",
        "description": "List all allowlist entries.",
        "operationId": "list_allowed_api_v1_admin_allowlist_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllowlistListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Add Allowed",
        "description": "Add a pattern to the allowlist.",
        "operationId": "add_allowed_api_v1_admin_allowlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllowlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Add Allowed Api V1 Admin Allowlist Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/allowlist/{pattern}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Remove Allowed",
        "description": "Remove a pattern from the allowlist.",
        "operationId": "remove_allowed_api_v1_admin_allowlist__pattern__delete",
        "parameters": [
          {
            "name": "pattern",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pattern"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove Allowed Api V1 Admin Allowlist  Pattern  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reputation/{address}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Reputation",
        "description": "Inspect agent reputation.",
        "operationId": "get_reputation_api_v1_admin_reputation__address__get",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Set Reputation",
        "description": "Admin override of reputation score.",
        "operationId": "set_reputation_api_v1_admin_reputation__address__put",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetReputationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReputationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/agents": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin List Agents",
        "description": "List all agents (including soft-deleted) for admin inspection.",
        "operationId": "admin_list_agents_api_v1_admin_agents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAgentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/agents/{address}/suspend": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Suspend Agent",
        "description": "Suspend an agent (admin action).",
        "operationId": "admin_suspend_agent_api_v1_admin_agents__address__suspend_post",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Audit Log",
        "description": "Query the audit log with optional filters.",
        "operationId": "admin_audit_log_api_v1_admin_audit_get",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminAgentListResponse": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AdminAgentResponse"
            },
            "type": "array",
            "title": "Agents"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "agents",
          "count"
        ],
        "title": "AdminAgentListResponse"
      },
      "AdminAgentResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "public_key": {
            "type": "string",
            "title": "Public Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "address",
          "public_key",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AdminAgentResponse"
      },
...
      "AllowlistEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pattern",
          "created_at"
        ],
        "title": "AllowlistEntry"
      },
      "AllowlistListResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AllowlistEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "count"
        ],
        "title": "AllowlistListResponse"
      },
      "AllowlistRequest": {
        "properties": {
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "pattern"
        ],
        "title": "AllowlistRequest"
      },
      "AuditLogEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "actor_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Address"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action",
          "entity_type",
          "entity_id",
          "timestamp"
        ],
        "title": "AuditLogEntry"
      },
      "AuditLogResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditLogEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "count"
        ],
        "title": "AuditLogResponse"
      },
      "BlocklistEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pattern",
          "created_at"
        ],
        "title": "BlocklistEntry"
      },
      "BlocklistListResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/BlocklistEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "count"
        ],
        "title": "BlocklistListResponse"
      },
      "BlocklistRequest": {
        "properties": {
          "pattern": {
            "type": "string",
            "title": "Pattern"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "pattern"
        ],
        "title": "BlocklistRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReputationResponse": {
        "properties": {
          "address": {
            "type": "string",
            "title": "Address"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "messages_sent": {
            "type": "integer",
            "title": "Messages Sent"
          },
          "messages_rejected": {
            "type": "integer",
            "title": "Messages Rejected"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "address",
          "score",
          "tier",
          "messages_sent",
          "messages_rejected",
          "created_at",
          "updated_at"
        ],
        "title": "ReputationResponse"
      },
...
      "SetReputationRequest": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "SetReputationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}